Compilation error fixes
- Compilation fixes.

Conflicts:
	src/java/android/telephony/SmsManager.java
	src/java/com/android/internal/telephony/PhoneBase.java
	src/java/com/android/internal/telephony/SubscriptionInfoUpdater.java
	src/java/com/android/internal/telephony/dataconnection/DcTracker.java
	src/java/com/android/internal/telephony/dataconnection/DcTrackerBase.java

Conflicts:
	src/java/com/android/internal/telephony/ModemStackController.java
	src/java/com/android/internal/telephony/cdma/CdmaServiceStateTracker.java
	src/java/com/android/internal/telephony/dataconnection/DctController.java

+++ src/java/android/telephony/CellBroadcastMessage.java
@@ -93,7 +93,7 @@ public class CellBroadcastMessage implements Parcelable {
         mSmsCbMessage = new SmsCbMessage(in);
         mDeliveryTime = in.readLong();
         mIsRead = (in.readInt() != 0);
-        mSubId = in.readLong();
+        mSubId = in.readInt();
     }
 
     /** Parcelable: no special flags. */
@@ -107,7 +107,7 @@ public class CellBroadcastMessage implements Parcelable {
         mSmsCbMessage.writeToParcel(out, flags);
         out.writeLong(mDeliveryTime);
         out.writeInt(mIsRead ? 1 : 0);
-        out.writeLong(mSubId);
+        out.writeInt(mSubId);
     }
 
     public static final Parcelable.Creator<CellBroadcastMessage> CREATOR
